+++ CHANGELOG.md
@@ -10,6 +10,18 @@ All notable changes to this project will be documented in this file.
 
 - Milesight - CT103
 
+# [1.73.0] - 2025-09-04
+
+### Added
+
+- Thermokon - SAB07
+
+# [1.72.0] - 2025-09-04
+
+### Added
+
+- Milesight - AM102
+
 # [1.71.0] - 2025-09-03
 
 ### Changed